修复

+++ src/FlightDisplay/RemoteControlSignalReceiver.h
@@ -20,7 +20,7 @@ class RemoteControlSignalReceiver : public QObject
 
     Q_PROPERTY(bool active READ active WRITE setActive NOTIFY activeChanged)
     Q_PROPERTY(bool listening READ listening NOTIFY listeningChanged)
-    Q_PROPERTY(quint16 port READ port WRITE setPort NOTIFY portChanged)
+    Q_PROPERTY(int port READ port WRITE setPort NOTIFY portChanged)
     Q_PROPERTY(QString lastPayload READ lastPayload NOTIFY lastPayloadChanged)
     Q_PROPERTY(QString statusText READ statusText NOTIFY statusTextChanged)
 
@@ -33,8 +33,8 @@ public:
 
     bool listening() const { return _listening; }
 
-    quint16 port() const { return _port; }
-    void setPort(quint16 port);
+    int port() const { return _port; }
+    void setPort(int port);
 
     QString lastPayload() const { return _lastPayload; }
     QString statusText() const { return _statusText; }
